Recipe search engine - REST API
Based on a previous project, I reused some code snippets and
also refactored some snippets to create this second version
of the "recipe finder".

Based on a previous project, I reused some code snippets and
also refactored some snippets to create this new version of
the "recipe search engine".

This commit includes the search engine being fully operational,
however unit tests and integration tests are not yet in place.

Also, the swagger spec does not exist.

This new version uses the hexagonal architecture pattern with a
Domain-Driven Design approach and seeks to optimize the performance
of the previous project.

issue search-engine-api